% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{bglmm}
\alias{bglmm}
\title{Bayesian Generalized Linear Mixed Model}
\usage{
bglmm(
  data,
  grp,
  grp_a = NULL,
  grp_b = NULL,
  id_var,
  x_var,
  y_vars,
  x_method = c("Empirical", "Analytical", "Value"),
  x_def = c(-Inf, Inf),
  test = c("right_sided", "left_sided"),
  rule = c("All", "Any", "Comp"),
  w = NULL,
  n_burn = 10000,
  n_it = 50000,
  start = c(0.5, 1),
  fixed = NULL,
  random = NULL,
  b_mu0 = NULL,
  b_sigma0 = NULL,
  g_mu0 = NULL,
  g_sigma0 = NULL,
  nu0 = NULL,
  tau0 = NULL,
  n_chain = 2,
  return_thinned = TRUE,
  n_thin = 10,
  return_diagnostics = TRUE,
  return_diagnostic_plots = FALSE,
  return_samples = FALSE
)
}
\arguments{
\item{data}{Data frame containing the data.}

\item{grp}{Character string. Name of the grouping variable.}

\item{grp_a}{Value of \code{grp} indicating first group (will be determined from factor levels if NULL).}

\item{grp_b}{Value of \code{grp} indicating second group (will be determined from factor levels if NULL).}

\item{id_var}{Character string. Name of cluster/ID variable.}

\item{x_var}{Character string. Name of covariate variable.}

\item{y_vars}{Character vector. Names of outcome variables (currently supports 2 outcomes).}

\item{x_method}{Character. Method for handling covariate. Default is "Empirical".}

\item{x_def}{Numeric. Defines subpopulation. Default is \code{c(-Inf, Inf)}.}

\item{test}{Character. Direction of test: "left_sided" for P(A>B) or "right_sided" for P(B>A).
Default is "right_sided".}

\item{rule}{Character. Decision rule: "All" (all outcomes favor hypothesis),
"Any" (at least one outcome favors hypothesis), or "Comp" (weighted combination).
Default is "All".}

\item{w}{Numeric vector. Weights for compensatory rule. Only used if rule = "Comp".
If NULL and rule = "Comp", equal weights are used. Default is NULL.}

\item{n_burn}{Integer. Number of burn-in iterations. Default is 10000.}

\item{n_it}{Integer. Number of MCMC iterations. Default is 50000 (takes long running time!).}

\item{start}{Numeric vector. Starting values for chains. Default is \code{c(0.5, 1)}.}

\item{fixed}{Character vector. Names of fixed effect variables. Default is c(x_var, grp_x_var).}

\item{random}{Character vector. Names of random effect variables. Default is c("Intercept", grp).}

\item{b_mu0}{Numeric vector. Prior means for fixed effects. Default is \code{rep(0, length(fixed))}.}

\item{b_sigma0}{Matrix. Prior covariance for fixed effects. Default is \code{diag(0.1, length(fixed))}.}

\item{g_mu0}{Numeric vector. Prior means for random effects. Default is \code{rep(0, length(random))}.}

\item{g_sigma0}{Matrix. Prior covariance for random effects. Default is \code{diag(0.1, length(random))}.}

\item{nu0}{Numeric. Prior degrees of freedom for inverse-Wishart. Default is \code{length(random)}.}

\item{tau0}{Matrix. Prior scale matrix of dimension \code{length(random) x length(random)} for inverse-Wishart. Default is \code{diag(1e-1, length(random))}.}

\item{n_chain}{Integer. Number of MCMC chains. Default is 2.}

\item{return_thinned}{Logical. Return thinned chains? Default is TRUE.}

\item{n_thin}{Integer. Thinning interval. Default is 10.}

\item{return_diagnostics}{Logical. Return MCMC diagnostics? Default is TRUE.}

\item{return_diagnostic_plots}{Logical. Should diagnostic plots (traceplots, autocorrelation, density) be returned? Default is \code{FALSE}. If \code{TRUE}, diagnostics are returned by default.}

\item{return_samples}{Logical. Return posterior samples? Default is FALSE.}
}
\value{
A list of class "bglmm".
}
\description{
Perform a Bayesian test for differences between two (sub)groups on multiple binary outcomes
using multilevel multinomial logistic regression.
}
