% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_rho}
\alias{compute_rho}
\title{Compute Pairwise Correlation}
\usage{
compute_rho(phi)
}
\arguments{
\item{phi}{Numeric vector of length 4 with joint response probabilities.}
}
\value{
Scalar. Correlation between outcome variables.
}
\description{
Compute pairwise correlation from joint response probabilities.
Based on Dai, Ding, & Wahba (2013). Multivariate Bernoulli distribution.
}
\keyword{internal}
