% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bglmm.R
\name{estimate_parameters_ml}
\alias{estimate_parameters_ml}
\title{Estimate Parameters for Multilevel Model}
\usage{
estimate_parameters_ml(
  X,
  Y,
  fixed,
  random,
  n_burn,
  n_it,
  start,
  b_mu0 = NULL,
  b_sigma0 = NULL,
  g_mu0 = NULL,
  g_sigma0 = NULL,
  nu0 = NULL,
  tau0 = NULL,
  n_chain = 2,
  return_thinned = FALSE,
  n_thin = 1
)
}
\arguments{
\item{X}{List of J (n_j x P) design matrices.}

\item{Y}{List of J (n_j x Q) response matrices.}

\item{fixed}{Character vector with names of fixed variables.}

\item{random}{Character vector with names of random variables.}

\item{n_burn}{Scalar. Number of burnin iterations.}

\item{n_it}{Scalar. Number of iterations.}

\item{start}{Vector of starting values for each chain.}

\item{b_mu0}{Vector of prior means of fixed regression coefficients.}

\item{b_sigma0}{Prior covariance matrix of fixed regression coefficients.}

\item{g_mu0}{Vector of prior means of random regression coefficients.}

\item{g_sigma0}{Prior covariance matrix of random regression coefficients.}

\item{nu0}{Scalar. Degrees of freedom of prior inverse-Wishart distribution.}

\item{tau0}{Prior matrix of inverse-Wishart distribution.}

\item{n_chain}{Scalar. Number of chains. Default is 2.}

\item{return_thinned}{Logical. Should thinned chains be returned? Default is FALSE.}

\item{n_thin}{Thinning rate. Default is 1.}
}
\value{
A named list with:
\describe{
\item{Pars}{List of parameter draws from each chain}
}
}
\description{
Estimate regression coefficients using multiple chains of Polya-Gamma Gibbs
sampling for multilevel data.
}
\keyword{internal}
