% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{integrand}
\alias{integrand}
\title{Integrand for Analytical Marginalization}
\usage{
integrand(
  x,
  est_rc,
  mu_x,
  sigma_x,
  range_x,
  grp,
  grp_var,
  population_var,
  q,
  fixed = NULL,
  random = NULL
)
}
\arguments{
\item{x}{Scalar. Value of covariate x.}

\item{est_rc}{(P x Q) matrix of regression coefficients.}

\item{mu_x}{Scalar. Mean of distribution of covariate.}

\item{sigma_x}{Scalar. Standard deviation of distribution of covariate.}

\item{range_x}{Numeric vector of lower and upper bound of range to integrate over.}

\item{grp}{Scalar. Value of group indicator (0 or 1).}

\item{q}{Scalar. Response category in 1 to Q.}

\item{fixed}{Optional character vector with names of fixed variables. Default is NULL.}

\item{random}{Optional character vector with names of random variables. Default is NULL.}
}
\value{
Scalar. Joint response probability for response category q.
}
\description{
Integrand function for integration over a range of a covariate in numerical
marginalization. Works for both GLM and GLMM.
}
\keyword{internal}
