% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multivariate2multinomial}
\alias{multivariate2multinomial}
\title{Transform Bivariate Binomial to Multinomial Response}
\usage{
multivariate2multinomial(y_bv)
}
\arguments{
\item{y_bv}{n x 2 matrix with bivariate binomial responses.}
}
\value{
n x 4 matrix with multinomial responses, ordered as \code{11}, \code{10}, \code{01}, \code{00}.
}
\description{
Transform bivariate binomial response data to multinomial response format.
}
\keyword{internal}
