% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{phi2theta}
\alias{phi2theta}
\title{Transform Phi to Theta}
\usage{
phi2theta(phi)
}
\arguments{
\item{phi}{Numeric vector of 4 joint response probabilities, summing to 1 and
ordered as \code{11}, \code{10}, \code{01}, \code{00}.}
}
\value{
Numeric vector of bivariate success probabilities.
}
\description{
Transform joint response probabilities to bivariate success probabilities.
Currently supported for K=2/Q=4 only.
}
\keyword{internal}
