% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bglm.R
\name{print.bglm}
\alias{print.bglm}
\title{Print Method for bglm Objects}
\usage{
\method{print}{bglm}(x, digits = 3, ...)
}
\arguments{
\item{x}{A bglm object.}

\item{digits}{Number of digits to display. Default is 3.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for bglm Objects
}
