% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bglmm.R
\name{print.bglmm}
\alias{print.bglmm}
\title{Print Method for bglmm Objects}
\usage{
\method{print}{bglmm}(x, digits = 3, ...)
}
\arguments{
\item{x}{A bglmm object.}

\item{digits}{Number of digits to display. Default is 3.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for bglmm Objects
}
\examples{
# Uses the pre-computed example object shipped with the package:
print(bglmm_fit)

}
\seealso{
\code{\link{summary.bglmm}}
}
