% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmvb.R
\name{print.bmvb}
\alias{print.bmvb}
\title{Print Method for bmvb Objects}
\usage{
\method{print}{bmvb}(x, digits = 3, ...)
}
\arguments{
\item{x}{A bmvb object.}

\item{digits}{Number of digits to display. Default is 3.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for bmvb Objects
}
