% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_mcmc.R
\name{print.mcmc_diagnostics}
\alias{print.mcmc_diagnostics}
\title{Print Method for mcmc_diagnostics Objects}
\usage{
\method{print}{mcmc_diagnostics}(x, ...)
}
\arguments{
\item{x}{A mcmc_diagnostics object.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for mcmc_diagnostics Objects
}
