% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plot_methods.R
\name{print.summary.bglm}
\alias{print.summary.bglm}
\title{Print Method for summary.bglm Objects}
\usage{
\method{print}{summary.bglm}(x, digits = 3, ...)
}
\arguments{
\item{x}{A summary.bglm object.}

\item{digits}{Number of digits. Default is 3.}

\item{...}{Additional arguments.}
}
\description{
Print Method for summary.bglm Objects
}
