% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plot_methods.R
\name{print.summary.bglmm}
\alias{print.summary.bglmm}
\title{Print Method for summary.bglmm Objects}
\usage{
\method{print}{summary.bglmm}(x, digits = 3, ...)
}
\arguments{
\item{x}{A summary.bglmm object.}

\item{digits}{Number of digits.}

\item{...}{Additional arguments.}
}
\description{
Print Method for summary.bglmm Objects
}
