% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bglmm.R
\name{sample_beta_ml}
\alias{sample_beta_ml}
\title{Sample Beta Coefficients for Multilevel Model}
\usage{
sample_beta_ml(
  X,
  Y,
  fixed,
  random,
  n_burn,
  n_it,
  start,
  b_mu0 = NULL,
  b_sigma0 = NULL,
  g_mu0 = NULL,
  g_sigma0 = NULL,
  nu0 = NULL,
  tau0 = NULL,
  return_thinned = FALSE,
  n_thin = 1
)
}
\arguments{
\item{X}{List of J (n_j x P) design matrices.}

\item{Y}{List of J (n_j x Q) response matrices.}

\item{fixed}{Character vector with names of fixed variables in covariate vector.}

\item{random}{Character vector with names of random variables in covariate vector.}

\item{n_burn}{Scalar. Number of burnin iterations.}

\item{n_it}{Scalar. Number of iterations.}

\item{start}{Vector of two starting values, each used for a different chain.}

\item{b_mu0}{Vector of prior means of fixed regression coefficients (length = no. of fixed covariates).}

\item{b_sigma0}{Prior precision matrix of fixed regression coefficients (length(fixed) x length(fixed)).}

\item{g_mu0}{Vector of prior means of random regression coefficients (length = no. of random covariates).}

\item{g_sigma0}{Prior precision matrix of random regression coefficients (length(random) x length(random)).}

\item{nu0}{Scalar. Degrees of freedom of prior inverse-Wishart distribution.}

\item{tau0}{Prior matrix of inverse-Wishart distribution.}

\item{return_thinned}{Logical. Should thinned chains be returned? Default is FALSE.}

\item{n_thin}{Thinning rate. Default is 1. Adjustable to integers > 1 when return_thinned = TRUE.}
}
\value{
A named list with:
\describe{
\item{b_draw_pg}{List of length n_it/n_thin with fixed regression coefficients (if fixed effects present)}
\item{g_draw_pg}{List of length n_it/n_thin with random regression coefficients (if random effects present)}
\item{gj_draw_pg}{List of length n_it/n_thin with cluster-specific random effects (if return_thinned = TRUE)}
\item{tau_draw_pg}{List of length n_it/n_thin with covariance matrices of random effects}
}
}
\description{
Sample regression coefficients via Polya-Gamma multinomial logistic regression
for multilevel data (single chain).
}
\keyword{internal}
