% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bglm.R
\name{sample_beta_pg}
\alias{sample_beta_pg}
\title{Sample Beta Coefficients using Polya-Gamma Method}
\usage{
sample_beta_pg(X, Y, n_burn, n_it, start, b_mu0, b_sigma0, verbose = FALSE)
}
\arguments{
\item{X}{(n x P) design matrix.}

\item{Y}{(n x Q) matrix of multinomial response data.}

\item{n_burn}{Scalar. Number of burnin iterations.}

\item{n_it}{Scalar. Number of iterations.}

\item{start}{Scalar. Starting value for this chain.}

\item{b_mu0}{Scalar. Prior mean.}

\item{b_sigma0}{Scalar. Prior precision of regression coefficients.}

\item{verbose}{Logical. If true, a progress bar is shown.}
}
\value{
b_draw_pg: List of length n_it with P x Q matrices with estimated regression coefficients
}
\description{
Sample regression coefficients via Polya-Gamma multinomial logistic regression
for a single chain.
}
\keyword{internal}
