% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_population}
\alias{sample_population}
\title{Sample Population Subsets}
\usage{
sample_population(
  X,
  population_var = NULL,
  grp_var,
  method,
  value = NULL,
  range = NULL,
  fixed = NULL,
  random = NULL
)
}
\arguments{
\item{X}{Design matrix with covariate data. Can be (n x P) matrix or list of J
(n_j x P) matrices for multilevel data.}

\item{population_var}{Optional character string with the variable name that defines
the subpopulation. Required when range is not c(-Inf, Inf).}

\item{grp_var}{Character string with name of variable that defines groups.}

\item{method}{Character. "Empirical" for empirical marginalization, "Value" for
vector of fixed values.}

\item{value}{Scalar. Value of x representing subpopulation. Required when method = "Value".}

\item{range}{Numeric vector of lower and upper bound of covariate that represents
subpopulation. Required when method = "Empirical".}

\item{fixed}{Character vector with names of fixed variables in covariate vector.
Default is NULL.}

\item{random}{Character vector with names of random variables in covariate vector.
Default is NULL.}
}
\value{
A list with:
\describe{
\item{x_a}{Design matrix with covariate data for group A}
\item{x_b}{Design matrix with covariate data for group B}
}
}
\description{
Extract subsets of covariate data for treatment groups based on specified method.
}
\keyword{internal}
