% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plot_methods.R
\name{summary.bglm}
\alias{summary.bglm}
\title{Summary Method for bglm Objects}
\usage{
\method{summary}{bglm}(object, prob = 0.95, ...)
}
\arguments{
\item{object}{A bglm object.}

\item{prob}{Numeric. Probability for credible intervals. Default is 0.95.}

\item{...}{Additional arguments.}
}
\value{
A summary.bglm object.
}
\description{
Summary Method for bglm Objects
}
