% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plot_methods.R
\name{summary.bglmm}
\alias{summary.bglmm}
\title{Summary Method for bglmm Objects}
\usage{
\method{summary}{bglmm}(object, prob = 0.95, ...)
}
\arguments{
\item{object}{A bglmm object.}

\item{prob}{Numeric. Probability for credible intervals.}

\item{...}{Additional arguments.}
}
\description{
Summary Method for bglmm Objects
}
