% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theta2delta_w}
\alias{theta2delta_w}
\title{Transform Theta to Weighted Treatment Difference}
\usage{
theta2delta_w(theta_a, theta_b, weights)
}
\arguments{
\item{theta_a}{Numeric vector of bivariate success probabilities for group A.}

\item{theta_b}{Numeric vector of bivariate success probabilities for group B.}

\item{weights}{Numeric vector of length K with weights for linear combination
of treatment differences (Compensatory rule).}
}
\value{
Scalar. Weighted treatment difference.
}
\description{
Transform bivariate success probabilities to weighted treatment difference.
Currently supported for K=2 only.
}
\keyword{internal}
