% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bglm.R
\name{transform2theta}
\alias{transform2theta}
\title{Transform Regression Coefficients to Success Probabilities}
\usage{
transform2theta(
  beta_draw_pg,
  X,
  Y = NULL,
  grp_var,
  population_var,
  measurement_level,
  method,
  range,
  value
)
}
\arguments{
\item{beta_draw_pg}{List of n_it (P x Q) matrices with posterior regression coefficients.}

\item{X}{(n x P) matrix with covariate data.}

\item{Y}{(n x Q) matrix with multinomial response data. Default is NULL.}

\item{measurement_level}{Character. "discrete" or "continuous".}

\item{method}{"Value" for vector of fixed values, "Empirical" for empirical
marginalization, "Analytical" for numerical marginalization.}

\item{range}{If method = "Analytical" or "Empirical" and if measurement_level is
continuous: range that defines the population of interest by a vector containing
a lower and an upper bound.}

\item{value}{If method = "Value": value that defines the population of interest
by a scalar value.}
}
\value{
A list with:
\describe{
\item{m_theta_a}{List of n_it vectors of length K with multivariate probabilities for group A}
\item{m_theta_b}{List of n_it vectors of length K with multivariate probabilities for group B}
}
}
\description{
Compute theta (success probabilities) from regression coefficients via various methods.
}
\keyword{internal}
