% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmco-package.R
\docType{package}
\name{bmco-package}
\alias{bmco}
\alias{bmco-package}
\title{bmco: Bayesian Analysis for Multivariate Categorical Outcomes}
\description{
Provides Bayesian methods for comparing groups on multiple binary outcomes,
including basic tests, regression adjustment, and multilevel models.
}
\section{Main functions}{

\itemize{
\item \code{\link{bmvb}}: Basic Bayesian test using multivariate Bernoulli
\item \code{\link{bglm}}: Test with covariate adjustment
\item \code{\link{bglmm}}: Multilevel/clustered data analysis
}
}

\author{
\strong{Maintainer}: Xynthia Kavelaars \email{xynthia.kavelaars@ou.nl} (\href{https://orcid.org/0000-0003-1600-3153}{ORCID})

}
